requirements.txt
matplotlib
tk
pyyaml
pymodbus